Infer code changes based on validFrom and validTo
We construct additional code changes when the `validTo` of a variant
matches the `validFrom` of the next variant. This is necessary because
name changes introduces new variants in the codes api, but not
necessarily new entries in the changes api. In order to connect variants
that aren't documented in the changes api, we therefore need to manually
check whether variants are adjacent.