Simplify interface to get_equivalents
The user now only needs to provide a classification ID and dates, and
optionally labeller functions and/or a graph object. The resulting
data.frame can be joined on to the user's own data. This also has the
added benefit of allowing us to use the dates provided by Klass
directly when performing the node search in `find_equivalent_nodes`,
instead of a potentially problematic date provided by the user.

The code has also been consolidated somewhat to reduce the number of
functions.