Add date controls to klass_graph
In cases where the date of input codes are known, it may
be useful to construct only part of the changes graph for a
given classification. We can use the `to` and `from` parameters
to control the time range we request from the Klass API.

This may be useful in cases where the Klass API struggles with
large requests, or when faster local performance is desired.